[Silver I] Title: 가장 가까운 세 사람의 심리적 거리, Time: 184 ms, Memory: 23804 KB -BaekjoonHub

+++ 백준/Silver/20529. 가장 가까운 세 사람의 심리적 거리/가장 가까운 세 사람의 심리적 거리.java
@@ -15,7 +15,7 @@ public class Main {
         StringBuilder sb = new StringBuilder();
 
         while (T-- > 0) {
-            long N = Long.parseLong(br.readLine());
+            int N = Integer.parseInt(br.readLine());
 
             String[] mbti = new String[(int) N];
